refactor(factory): build @ttsc/factory with tsgo + shared rollup config
Drop the `ttsc` build dependency — @ttsc/factory has no plugins, so plain
`tsgo` compiles it. Keep the dual CJS/ESM publish (`tsgo` emits the CJS
`lib/*.js`, rollup emits the ESM `lib/*.mjs`) by adopting the typia-style
shared rollup config: `config/rollup.config.mjs` holds the config and
`@ttsc/config` owns the rollup devDependencies; each package's
`rollup.config.mjs` just re-exports it.

The package compiles with `module: nodenext` (like typia) so rollup's
commonjs→esm step produces a valid `lib/index.mjs` default export.

Restructure the node factory for tree-shakeable bundles: `src/factory/index.ts`
is now a pure `export *` re-export barrel of the per-node builders (the
hand-assembled object literal in the former `factory.ts` is gone), and
`src/index.ts` aggregates them into the `factory` namespace via
`export * as factory from "./factory/index"`. The default export and
`NodeFactory` type are preserved, so `import factory, { TsPrinter } from
"@ttsc/factory"` keeps working.
