+++ packages/factory/rollup.config.mjs
@@ -1,31 +1,1 @@
-import commonjs from "@rollup/plugin-commonjs";
-import nodeResolve from "@rollup/plugin-node-resolve";
-import autoExternal from "rollup-plugin-auto-external";
-import nodeExternals from "rollup-plugin-node-externals";
-import { globSync } from "tinyglobby";
-
-export default {
-  input: globSync("./lib/**/*.js"),
-  external: (id) => /node_modules/.test(id),
-  output: {
-    dir: "./lib",
-    format: "esm",
-    sourcemap: true,
-    entryFileNames: (chunkInfo) => {
-      if (chunkInfo.name.includes("node_modules")) {
-        throw new Error(`Invalid chunk name: ${chunkInfo.name}`);
-      }
-      return `[name].mjs`;
-    },
-    preserveModules: true,
-    preserveModulesRoot: "lib",
-  },
-  plugins: [
-    nodeExternals(),
-    autoExternal(),
-    nodeResolve(),
-    commonjs({
-      strictRequires: false,
-    }),
-  ],
-};
+export { default } from "../../config/rollup.config.mjs";
