+++ packages/factory/src/factory/index.ts
@@ -1,2 +1,198 @@
-export { factory } from "./factory";
-export type { NodeFactory } from "./factory";
+/**
+ * Hand-written, dependency-free re-implementation of the legacy TypeScript AST
+ * node factory (`ts.factory`).
+ *
+ * This module is a pure re-export barrel: each `createXxx` builder lives in its
+ * own file and is re-exported here. `src/index.ts` aggregates them into the
+ * `factory` namespace via `export * as factory`, which lets bundlers tree-shake
+ * unused builders. No `typescript` module is imported anywhere.
+ *
+ * @author Jeongho Nam - https://github.com/samchon
+ */
+export * from "./clauses/createHeritageClause";
+export * from "./clauses/createParameterDeclaration";
+export * from "./declarations/createClassDeclaration";
+export * from "./declarations/createClassStaticBlockDeclaration";
+export * from "./declarations/createConstructorDeclaration";
+export * from "./declarations/createEnumDeclaration";
+export * from "./declarations/createEnumMember";
+export * from "./declarations/createExternalModuleReference";
+export * from "./declarations/createFunctionDeclaration";
+export * from "./declarations/createGetAccessorDeclaration";
+export * from "./declarations/createImportEqualsDeclaration";
+export * from "./declarations/createInterfaceDeclaration";
+export * from "./declarations/createMethodDeclaration";
+export * from "./declarations/createModuleBlock";
+export * from "./declarations/createModuleDeclaration";
+export * from "./declarations/createNamespaceExportDeclaration";
+export * from "./declarations/createPropertyDeclaration";
+export * from "./declarations/createSemicolonClassElement";
+export * from "./declarations/createSetAccessorDeclaration";
+export * from "./declarations/createTypeAliasDeclaration";
+export * from "./expressions/createAdd";
+export * from "./expressions/createArrayBindingPattern";
+export * from "./expressions/createArrayLiteralExpression";
+export * from "./expressions/createArrowFunction";
+export * from "./expressions/createAsExpression";
+export * from "./expressions/createAssignment";
+export * from "./expressions/createAwaitExpression";
+export * from "./expressions/createBinaryExpression";
+export * from "./expressions/createBindingElement";
+export * from "./expressions/createBitwiseAnd";
+export * from "./expressions/createBitwiseNot";
+export * from "./expressions/createBitwiseOr";
+export * from "./expressions/createBitwiseXor";
+export * from "./expressions/createCallChain";
+export * from "./expressions/createCallExpression";
+export * from "./expressions/createClassExpression";
+export * from "./expressions/createComma";
+export * from "./expressions/createCommaListExpression";
+export * from "./expressions/createComputedPropertyName";
+export * from "./expressions/createConditionalExpression";
+export * from "./expressions/createDeleteExpression";
+export * from "./expressions/createDivide";
+export * from "./expressions/createElementAccessChain";
+export * from "./expressions/createElementAccessExpression";
+export * from "./expressions/createEquality";
+export * from "./expressions/createExponent";
+export * from "./expressions/createFunctionExpression";
+export * from "./expressions/createGreaterThan";
+export * from "./expressions/createGreaterThanEquals";
+export * from "./expressions/createImmediatelyInvokedArrowFunction";
+export * from "./expressions/createImmediatelyInvokedFunctionExpression";
+export * from "./expressions/createInequality";
+export * from "./expressions/createLeftShift";
+export * from "./expressions/createLessThan";
+export * from "./expressions/createLessThanEquals";
+export * from "./expressions/createLogicalAnd";
+export * from "./expressions/createLogicalNot";
+export * from "./expressions/createLogicalOr";
+export * from "./expressions/createMetaProperty";
+export * from "./expressions/createModulo";
+export * from "./expressions/createMultiply";
+export * from "./expressions/createNewExpression";
+export * from "./expressions/createNonNullChain";
+export * from "./expressions/createNonNullExpression";
+export * from "./expressions/createObjectBindingPattern";
+export * from "./expressions/createObjectLiteralExpression";
+export * from "./expressions/createOmittedExpression";
+export * from "./expressions/createParenthesizedExpression";
+export * from "./expressions/createPostfixDecrement";
+export * from "./expressions/createPostfixIncrement";
+export * from "./expressions/createPostfixUnaryExpression";
+export * from "./expressions/createPrefixDecrement";
+export * from "./expressions/createPrefixIncrement";
+export * from "./expressions/createPrefixMinus";
+export * from "./expressions/createPrefixPlus";
+export * from "./expressions/createPrefixUnaryExpression";
+export * from "./expressions/createPropertyAccessChain";
+export * from "./expressions/createPropertyAccessExpression";
+export * from "./expressions/createPropertyAssignment";
+export * from "./expressions/createRegularExpressionLiteral";
+export * from "./expressions/createRightShift";
+export * from "./expressions/createSatisfiesExpression";
+export * from "./expressions/createShorthandPropertyAssignment";
+export * from "./expressions/createSpreadAssignment";
+export * from "./expressions/createSpreadElement";
+export * from "./expressions/createStrictEquality";
+export * from "./expressions/createStrictInequality";
+export * from "./expressions/createSubtract";
+export * from "./expressions/createTaggedTemplateExpression";
+export * from "./expressions/createTemplateExpression";
+export * from "./expressions/createTemplateSpan";
+export * from "./expressions/createTypeAssertion";
+export * from "./expressions/createTypeOfExpression";
+export * from "./expressions/createUnsignedRightShift";
+export * from "./expressions/createVoidExpression";
+export * from "./expressions/createVoidZero";
+export * from "./expressions/createYieldExpression";
+export * from "./file/createNodeArray";
+export * from "./file/createSourceFile";
+export * from "./file/updateSourceFile";
+export * from "./imports/createExportAssignment";
+export * from "./imports/createExportDeclaration";
+export * from "./imports/createExportDefault";
+export * from "./imports/createExportSpecifier";
+export * from "./imports/createExternalModuleExport";
+export * from "./imports/createImportClause";
+export * from "./imports/createImportDeclaration";
+export * from "./imports/createImportSpecifier";
+export * from "./imports/createNamedExports";
+export * from "./imports/createNamedImports";
+export * from "./imports/createNamespaceExport";
+export * from "./imports/createNamespaceImport";
+export * from "./literals/createBigIntLiteral";
+export * from "./literals/createNoSubstitutionTemplateLiteral";
+export * from "./literals/createNumericLiteral";
+export * from "./literals/createStringLiteral";
+export * from "./literals/createTemplateHead";
+export * from "./literals/createTemplateMiddle";
+export * from "./literals/createTemplateTail";
+export * from "./names/createDecorator";
+export * from "./names/createFalse";
+export * from "./names/createIdentifier";
+export * from "./names/createModifier";
+export * from "./names/createNull";
+export * from "./names/createPrivateIdentifier";
+export * from "./names/createQualifiedName";
+export * from "./names/createSuper";
+export * from "./names/createThis";
+export * from "./names/createToken";
+export * from "./names/createTrue";
+export * from "./statements/createBlock";
+export * from "./statements/createBreakStatement";
+export * from "./statements/createCaseBlock";
+export * from "./statements/createCaseClause";
+export * from "./statements/createCatchClause";
+export * from "./statements/createContinueStatement";
+export * from "./statements/createDebuggerStatement";
+export * from "./statements/createDefaultClause";
+export * from "./statements/createDoStatement";
+export * from "./statements/createEmptyStatement";
+export * from "./statements/createExpressionStatement";
+export * from "./statements/createForInStatement";
+export * from "./statements/createForOfStatement";
+export * from "./statements/createForStatement";
+export * from "./statements/createIfStatement";
+export * from "./statements/createLabeledStatement";
+export * from "./statements/createReturnStatement";
+export * from "./statements/createSwitchStatement";
+export * from "./statements/createThrowStatement";
+export * from "./statements/createTryStatement";
+export * from "./statements/createVariableDeclaration";
+export * from "./statements/createVariableDeclarationList";
+export * from "./statements/createVariableStatement";
+export * from "./statements/createWhileStatement";
+export * from "./statements/createWithStatement";
+export * from "./types/createArrayTypeNode";
+export * from "./types/createCallSignature";
+export * from "./types/createConditionalTypeNode";
+export * from "./types/createConstructorTypeNode";
+export * from "./types/createConstructSignature";
+export * from "./types/createExpressionWithTypeArguments";
+export * from "./types/createFunctionTypeNode";
+export * from "./types/createImportTypeNode";
+export * from "./types/createIndexedAccessTypeNode";
+export * from "./types/createIndexSignature";
+export * from "./types/createInferTypeNode";
+export * from "./types/createIntersectionTypeNode";
+export * from "./types/createKeywordTypeNode";
+export * from "./types/createLiteralTypeNode";
+export * from "./types/createMappedTypeNode";
+export * from "./types/createMethodSignature";
+export * from "./types/createNamedTupleMember";
+export * from "./types/createOptionalTypeNode";
+export * from "./types/createParenthesizedType";
+export * from "./types/createPropertySignature";
+export * from "./types/createRestTypeNode";
+export * from "./types/createTemplateLiteralType";
+export * from "./types/createTemplateLiteralTypeSpan";
+export * from "./types/createThisTypeNode";
+export * from "./types/createTupleTypeNode";
+export * from "./types/createTypeLiteralNode";
+export * from "./types/createTypeOperatorNode";
+export * from "./types/createTypeParameterDeclaration";
+export * from "./types/createTypePredicateNode";
+export * from "./types/createTypeQueryNode";
+export * from "./types/createTypeReferenceNode";
+export * from "./types/createUnionTypeNode";
