+++ packages/factory/tsconfig.json
@@ -1,6 +1,8 @@
 {
   "extends": "../../config/tsconfig.json",
   "compilerOptions": {
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
     "rootDir": "src",
     "outDir": "lib"
   },
